michael-mathew_123140101_praktikum6/requirements.txt
pyramid==2.0.2
waitress==2.1.2
sqlalchemy==1.4.53
alembic==1.13.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
zope.sqlalchemy==2.0
plaster-pastedeploy==0.7
pyramid-tm==2.4
transaction==3.1.0

